chore: refactor components #110
